refactor: enable MQTT and logger initialization, update user role check, and enhance email sending with retry logic

+++ be/controllers/users.controllers.js
@@ -154,7 +154,7 @@ const adminKeyPin = async (req, res) => {
 const deleteroom = async (req, res) => {
     if (process.env.DEBUG_MODE === "false") {
         const except_rooms = (process.env.EXECPT_ROOMS || "").split(",").map(num => Number(num.trim()));
-        const { eventId, room_number } = req.body; // , eventId
+        const { eventId, room_number } = req.body; // eventId, room_number
         if (except_rooms.includes(Number(room_number))) {
             return res.status(200).json({ message: `Room ${room_number} is in the exception list, skip delete.` });
         }
@@ -163,23 +163,34 @@ const deleteroom = async (req, res) => {
             return res.status(401).json({ error: "No refresh token found. Please login again." });
         }
         try {
+            const eventRecord = await waitUntil(async () => {
+                const try_eventRecord = await bookingkey.findOne(
+                    { eventId },
+                    { isPinVerified: 1, organizerMail: 1, room: 1 }
+                );
+                // ถ้าหาข้อมูลเจอ แล้วถึงจะ return ไม่งั้นก็วนรอไปก่อน safety timeout
+                if (try_eventRecord) return try_eventRecord;
+                return null;
+            }, 10000, 2000);
+
+            if (!eventRecord) {
+                return res.status(404).json({ message: "Event not found in DB" });
+            }
+
+            // ถ้า verify แล้ว หรือลบเเล้ว → ไม่ต้องลบ
+            if (eventRecord.isPinVerified === true || eventRecord.isPinVerified === "true" || eventRecord.isPinVerified === "not access") {
+                return res.status(200).json({
+                    message: `Event ${eventId} already verified by PIN or was deleted. Skip delete.`,
+                });
+            }
+            
             await getGraphClient(AccessToken)
                 .api(`/me/events/${eventId}`)
                 .delete();
 
-            console.log("Delete event success");
-            const falselist = await waitUntil(async () => {
-                const result = await bookingkey.findOneAndUpdate(
-                    { eventId },
-                    { isPinVerified: 'not access' },
-                    { new: true }
-                );
-                console.log("[WaitUntil] result:", result);
-                // ถ้ามี organizerMail แล้วถึงจะ return
-                if (result?.organizerMail) return result;
-                return null;
-            }, 30000, 3000);
-            const organizerEmail = falselist.organizerMail;
+            eventRecord.isPinVerified = "not access";
+            await eventRecord.save();   
+            const organizerEmail = eventRecord.organizerMail;
 
             if (organizerEmail && organizerEmail !== process.env.CENTERLIZED_MAIL) {
                 // ใช้ upsert เพื่อสร้างใหม่หากไม่มี หรือ update หากมีอยู่แล้ว
@@ -191,7 +202,7 @@ const deleteroom = async (req, res) => {
                             EventId: {
                                 eventId,
                                 missedAt: new Date(), //await GetTimeAPI('Asia/Bangkok'),
-                                RoomNumber: Number(falselist.room)
+                                RoomNumber: Number(eventRecord.room)
                             }
                         }
                     },
